Keep archived comment sanitizer server-side
