+++ src/routes/diary/[slug]/+page.ts
@@ -1,24 +1,12 @@
 import { type PostMetadata, type PostSummary } from '../../api/posts/util';
-import { unwrapArchivedComments } from './comments';
 import type { PageLoad } from './$types';
 
-const archivedCommentFiles = import.meta.glob('/src/lib/diary/**/comments.md', {
-  query: '?raw',
-  import: 'default',
-});
-
-export const load: PageLoad = async ({ fetch, params }) => {
+export const load: PageLoad = async ({ data, fetch, params }) => {
   const response = await fetch(`/api/posts/all`);
   const posts = (await response.json()) as PostSummary[];
 
   const { slug } = params;
   const post = await import(`../../../lib/diary/${slug}/page.md`);
-  const archivedCommentsLoader = archivedCommentFiles[`/src/lib/diary/${slug}/comments.md`] as
-    | (() => Promise<string>)
-    | undefined;
-  const archivedCommentsHtml = archivedCommentsLoader
-    ? unwrapArchivedComments(await archivedCommentsLoader())
-    : undefined;
   const metadata = post.metadata as PostMetadata;
   const { title, date } = metadata;
   const PostComponent = post.default;
@@ -37,7 +25,7 @@ export const load: PageLoad = async ({ fetch, params }) => {
       title,
       date,
       PostComponent,
-      archivedCommentsHtml,
+      archivedCommentsHtml: data.archivedCommentsHtml,
       metadata,
     },
   };
